Refactoring in order to improve usability
- Success/failure folder structure is removed from the requirements.
Now, status codes should be inserted into the filenames. One exception
is that {httpMethod}.{encoding} will be considered as
{httpMethod}_200.{encoding} by default and have higher precedence.
- Mock suites can be initialized by setting base url or bundle name.
This will make it possible to support multiple base urls with related bundles.
- Mock bundle versioning is simply provided for the same base url. This way,
setting a list of bundle names, you can override paths so that the last bundle
in the list will have the highest precedence while returning the mock.
- Example project is tweaked a bit to be able to test newly-developed features.
- Some changes on the code styling.